refactor: move the Address and UDT types to src/models
