+++ src/models/Address/index.ts
@@ -1,0 +1,31 @@
+import { Script } from '../Script'
+import { UDTAccount } from './UDTAccount'
+
+export interface LockInfo {
+  status: 'locked' | 'unlocked'
+  epochNumber: string
+  epochIndex: string
+  estimatedUnlockTime: string
+}
+
+export interface Address {
+  addressHash: string
+  lockHash: string
+  balance: string
+  balanceOccupied: string
+  transactionsCount: number
+  lockScript: Script
+  pendingRewardBlocksCount: number
+  type: 'Address' | 'LockHash' | ''
+  daoDeposit: number
+  interest: number
+  daoCompensation: number
+  lockInfo: LockInfo
+  liveCellsCount: string
+  minedBlocksCount: string
+  isSpecial: boolean
+  specialAddress: string
+  udtAccounts?: Array<UDTAccount>
+}
+
+export * from './UDTAccount'
